requirements.txt
git+https://github.com/openai/swarm.git
openai==1.55.3
langchain==0.1.20
faiss-cpu
pypdf==5.1.0
python-dotenv==1.0.0
reportlab==4.1.0
httpx==0.27.2
tiktoken==0.6.0
langchain-openai>=0.0.2
ipywidgets>=8.0.0
streamlit==1.30.0
PyMuPDF==1.23.8
